fixes + bd nueva

+++ static/js/alumnos.js
@@ -110,7 +110,9 @@
                  );
          };
          $scope.eliminar = function(id) {
-             $http({
+
+             if (confirm("¿Está seguro que desea eliminarlo?"))
+                 $http({
                      method: 'DELETE',
                      url: '/escuela/alumnos/' + id + '/'
                  })
