+++ static/js/cursos.js
@@ -93,7 +93,9 @@
                  );
          };
          $scope.eliminar = function(id) {
-             $http({
+
+             if (confirm("¿Está seguro que desea eliminarlo?"))
+                 $http({
                      method: 'DELETE',
                      url: '/escuela/cursos/' + id + '/'
                  })
